+++ flodl/src/distributed/mod.rs
@@ -22,7 +22,7 @@ pub mod el_che;
 pub use cuda_event::{CudaEvent, CudaEventFlags};
 pub use cuda_stream::{CudaStream, StreamGuard};
 pub use nccl::{NcclAbortHandle, NcclComms, NcclRankComm, NcclUniqueId, ReduceOp};
-pub use ddp::{Ddp, DdpConfig, Trainer};
+pub use ddp::{Ddp, DdpConfig, HasGraph, Trainer};
 pub use el_che::ElChe;
 pub use ddp_run::{ApplyPolicy, DdpHandle, DdpBuilder, DdpRunConfig, AverageBackend, TrainedState, EpochMetrics, record_scalar, drain_scalars, GpuWorker};
 // Deprecated aliases